Inline textContains() helper function
