+++ lib/assertions.js
@@ -1,7 +1,6 @@
 import exists from './assertions/exists';
 import focused from './assertions/focused';
 import notFocused from './assertions/not-focused';
-import textContains from './assertions/text-contains';
 import textMatches from './assertions/text-matches';
 
 const elementToString = require('./helpers/element-to-string');
@@ -78,7 +77,18 @@ export default class DOMAssertions {
    * assert.dom('#title').textContains('Welcome');
    */
   textContains(text, message) {
-    textContains.call(this, text, message);
+    let element = this.findTargetElement();
+    if (!element) return;
+
+    let result = element.textContent.indexOf(text) !== -1;
+    let actual = element.textContent;
+    let expected = text;
+
+    if (!message) {
+      message = `Element ${this.targetDescription} contains "${text}"`;
+    }
+
+    this.pushResult({ result, actual, expected, message });
   }
 
   /**
